+++ spec/slack/smart-bot/commands/general/general_commands/teams/add_team_spec.rb
