+++ spec/slack/smart-bot/commands/general/general_commands/teams/delete_team_spec.rb
