+++ spec/slack/smart-bot/commands/general/general_commands/teams/memos/delete_memo_team_spec.rb
