+++ spec/slack/smart-bot/commands/general/general_commands/teams/memos/see_memo_team_spec.rb
