+++ spec/slack/smart-bot/commands/general/general_commands/teams/ping_team_spec.rb
