+++ spec/slack/smart-bot/commands/general/general_commands/teams/see_teams_spec.rb
