+++ spec/slack/smart-bot/commands/general/general_commands/teams/see_vacations_team_spec.rb
