+++ spec/slack/smart-bot/commands/general/general_commands/teams/update_team_spec.rb
